Generate image thumbnails in the legacy composer via a shared chat-image-preview helper
