+++ apps/ui/src/ui-classic/components/session-view/composer/index.tsx
@@ -1,11 +1,13 @@
 import {
+	createStudioChatImagePreview,
+	loadImageElement,
+} from '@studio/common/ai/chat-image-preview';
+import {
 	STUDIO_CHAT_IMAGE_MIME_TYPES,
 	STUDIO_CHAT_MAX_IMAGES,
 	STUDIO_CHAT_MAX_IMAGE_BYTES,
-	STUDIO_CHAT_MAX_IMAGE_PREVIEW_BYTES,
 	isStudioChatImageMimeType,
 	validateStudioChatImages,
-	type StudioChatImageMimeType,
 } from '@studio/common/ai/chat-images';
 import { AI_MODELS, getAiModelFamily, getAiModelLabel } from '@studio/common/ai/models';
 import { isStudioCustomEntryOfType } from '@studio/common/ai/sessions/entry-types';
@@ -62,53 +64,6 @@ function readFileAsDataUrl( file: File ): Promise< string > {
 		reader.onerror = () => reject( new Error( __( 'Unable to read the selected image.' ) ) );
 		reader.readAsDataURL( file );
 	} );
-}
-
-function loadImageElement( src: string ): Promise< HTMLImageElement | undefined > {
-	return new Promise( ( resolve ) => {
-		const image = new Image();
-		image.onload = () => resolve( image );
-		image.onerror = () => resolve( undefined );
-		image.src = src;
-	} );
-}
-
-const PREVIEW_MAX_DIMENSION_PX = 256;
-const PREVIEW_JPEG_QUALITY = 0.8;
-
-// Downscales the attached image into a small thumbnail for transcript chips.
-// The thumbnail is what gets persisted on the display entry — the full bytes
-// only live in the model's message — so it must stay genuinely small.
-function createPreviewDataUrl(
-	image: HTMLImageElement,
-	mimeType: StudioChatImageMimeType
-): string | undefined {
-	const scale = Math.min(
-		1,
-		PREVIEW_MAX_DIMENSION_PX / Math.max( image.naturalWidth, image.naturalHeight, 1 )
-	);
-	const width = Math.max( 1, Math.round( image.naturalWidth * scale ) );
-	const height = Math.max( 1, Math.round( image.naturalHeight * scale ) );
-	const canvas = document.createElement( 'canvas' );
-	canvas.width = width;
-	canvas.height = height;
-	const context = canvas.getContext( '2d' );
-	if ( ! context ) {
-		return undefined;
-	}
-	context.drawImage( image, 0, 0, width, height );
-	try {
-		// JPEG drops the alpha channel, so only JPEG sources re-encode as JPEG.
-		const previewDataUrl =
-			mimeType === 'image/jpeg'
-				? canvas.toDataURL( 'image/jpeg', PREVIEW_JPEG_QUALITY )
-				: canvas.toDataURL( 'image/png' );
-		return previewDataUrl.length <= STUDIO_CHAT_MAX_IMAGE_PREVIEW_BYTES
-			? previewDataUrl
-			: undefined;
-	} catch {
-		return undefined;
-	}
 }
 
 function fileListToArray( files: FileList | null | undefined ) {
@@ -264,7 +219,7 @@ export const Composer = forwardRef< ComposerHandle, ComposerProps >( function Co
 				size: file.size,
 				width: imageElement?.naturalWidth,
 				height: imageElement?.naturalHeight,
-				previewDataUrl: imageElement && createPreviewDataUrl( imageElement, mimeType ),
+				previewDataUrl: imageElement && createStudioChatImagePreview( imageElement, mimeType ),
 				dataBase64,
 				previewUrl,
 			};
